Geocode Single Coordinate Pair

Provides access to the GeoLookup API of the US Census Bureau. Returns census geographies for a single geographic point.

cxy_geography(
  lon,
  lat,
  benchmark = "Public_AR_Current",
  vintage = "Current_Current"
)

Arguments

lon

Numeric or String Containing Longitude (x) of Point

lat

Numeric or String Containing Latitude (y) of Point

benchmark

Optional ID or Name of Census Benchmark. See Details.

vintage

Optional ID or Name of Census Vintage. See Details.

Value

A data.frame containing matched address or NULL if not matches

Details

This function can be used to locate geographic information given a geographic point. It does not provide an address like a reverse-geocoder

To obtain current valid benchmarks, use the cxy_benchmarks() function

To use this function, you must specify a valid vintage for your benchmark. You may use the cxy_vintages() function to obtain valid Vintages for a given benchmark. See vignette('censusxy') for a full walkthrough.

Examples

# \donttest{
cxy_geography(lon = -90.23324, lat = 38.63593)
#>   States.STATENS States.GEOID States.CENTLAT States.AREAWATER States.STATE
#> 1       01779791           29    +38.3569829       2487375487           29
#>   States.BASENAME States.STUSAB    States.OID States.LSADC States.FUNCSTAT
#> 1        Missouri            MO 2749046934061           00               A
#>   States.INTPTLAT States.DIVISION States.NAME States.REGION States.OBJECTID
#> 1     +38.3507500               4    Missouri             2               4
#>   States.CENTLON States.AREALAND States.INTPTLON States.MTFCC
#> 1   -092.4580564    178052403953    -092.4567826        G4000
#>   Combined.Statistical.Areas.POP100 Combined.Statistical.Areas.GEOID
#> 1                                                                476
#>   Combined.Statistical.Areas.CENTLAT Combined.Statistical.Areas.AREAWATER
#> 1                        +38.6839906                            582100220
#>       Combined.Statistical.Areas.BASENAME Combined.Statistical.Areas.OID
#> 1 St. Louis-St. Charles-Farmington, MO-IL               2619034687908121
#>   Combined.Statistical.Areas.LSADC Combined.Statistical.Areas.FUNCSTAT
#> 1                               M0                                   S
#>   Combined.Statistical.Areas.INTPTLAT
#> 1                         +38.6843841
#>               Combined.Statistical.Areas.NAME
#> 1 St. Louis-St. Charles-Farmington, MO-IL CSA
#>   Combined.Statistical.Areas.OBJECTID Combined.Statistical.Areas.CSA
#> 1                                  63                            476
#>   Combined.Statistical.Areas.CENTLON Combined.Statistical.Areas.INTPTLON
#> 1                       -090.2663583                        -090.2644483
#>   Combined.Statistical.Areas.AREALAND Combined.Statistical.Areas.HU100
#> 1                         23019677209                                 
#>   Combined.Statistical.Areas.MTFCC County.Subdivisions.COUSUB
#> 1                            G3100                      65000
#>   County.Subdivisions.GEOID County.Subdivisions.CENTLAT
#> 1                2951065000                 +38.6357916
#>   County.Subdivisions.AREAWATER County.Subdivisions.STATE
#> 1                      11534970                        29
#>   County.Subdivisions.BASENAME County.Subdivisions.OID
#> 1                    St. Louis          27690325436609
#>   County.Subdivisions.LSADC County.Subdivisions.FUNCSTAT
#> 1                        25                            F
#>   County.Subdivisions.INTPTLAT County.Subdivisions.NAME
#> 1                  +38.6356988           St. Louis city
#>   County.Subdivisions.OBJECTID County.Subdivisions.CENTLON
#> 1                        20434                -090.2451153
#>   County.Subdivisions.COUSUBCC County.Subdivisions.AREALAND
#> 1                           Z7                    159853177
#>   County.Subdivisions.INTPTLON County.Subdivisions.MTFCC
#> 1                 -090.2445816                     G4040
#>   County.Subdivisions.COUSUBNS County.Subdivisions.COUNTY Urban.Areas.GEOID
#> 1                     00767557                        510             77770
#>   Urban.Areas.CENTLAT Urban.Areas.AREAWATER Urban.Areas.BASENAME
#> 1         +38.6375942              32000217    St. Louis, MO--IL
#>   Urban.Areas.OID Urban.Areas.UA Urban.Areas.LSADC Urban.Areas.FUNCSTAT
#> 1  27021149986311          77770                67                    S
#>   Urban.Areas.INTPTLAT             Urban.Areas.NAME Urban.Areas.OBJECTID
#> 1          +38.6363760 St. Louis, MO--IL Urban Area                 1309
#>   Urban.Areas.CENTLON Urban.Areas.AREALAND Urban.Areas.INTPTLON
#> 1        -090.3423874           2357817363         -090.3414179
#>   Urban.Areas.MTFCC Incorporated.Places.DISP_CLR Incorporated.Places.NECTAPCI
#> 1             G3500                            2                            N
#>   Incorporated.Places.GEOID Incorporated.Places.CENTLAT
#> 1                   2965000                 +38.6357916
#>   Incorporated.Places.AREAWATER Incorporated.Places.BASENAME
#> 1                      11534970                    St. Louis
#>   Incorporated.Places.STATE Incorporated.Places.OID Incorporated.Places.LSADC
#> 1                        29          27890325442361                        25
#>   Incorporated.Places.PLACE Incorporated.Places.FUNCSTAT
#> 1                     65000                            A
#>   Incorporated.Places.INTPTLAT Incorporated.Places.NAME
#> 1                  +38.6356988           St. Louis city
#>   Incorporated.Places.OBJECTID Incorporated.Places.PLACECC
#> 1                         9092                          C7
#>   Incorporated.Places.CENTLON Incorporated.Places.CBSAPCI
#> 1                -090.2451153                           Y
#>   Incorporated.Places.AREALAND Incorporated.Places.INTPTLON
#> 1                    159853177                 -090.2445816
#>   Incorporated.Places.PLACENS Incorporated.Places.MTFCC Counties.GEOID
#> 1                    00767557                     G4110          29510
#>   Counties.CENTLAT Counties.AREAWATER Counties.STATE Counties.BASENAME
#> 1      +38.6357916           11534970             29         St. Louis
#>     Counties.OID Counties.LSADC Counties.FUNCSTAT Counties.INTPTLAT
#> 1 27590325427938             25                 F       +38.6356988
#>    Counties.NAME Counties.OBJECTID Counties.CENTLON Counties.COUNTYCC
#> 1 St. Louis city              2198     -090.2451153                C7
#>   Counties.COUNTYNS Counties.AREALAND Counties.INTPTLON Counties.MTFCC
#> 1          00767557         159853177      -090.2445816          G4020
#>   Counties.COUNTY X2024.State.Legislative.Districts...Upper.GEOID
#> 1             510                                           29005
#>   X2024.State.Legislative.Districts...Upper.CENTLAT
#> 1                                       +38.6496700
#>   X2024.State.Legislative.Districts...Upper.AREAWATER
#> 1                                            11253685
#>   X2024.State.Legislative.Districts...Upper.STATE
#> 1                                              29
#>   X2024.State.Legislative.Districts...Upper.BASENAME
#> 1                                                  5
#>   X2024.State.Legislative.Districts...Upper.OID
#> 1                              2129035951821972
#>   X2024.State.Legislative.Districts...Upper.SLDU
#> 1                                            005
#>   X2024.State.Legislative.Districts...Upper.LSADC
#> 1                                              LU
#>   X2024.State.Legislative.Districts...Upper.FUNCSTAT
#> 1                                                  N
#>   X2024.State.Legislative.Districts...Upper.INTPTLAT
#> 1                                        +38.6497548
#>   X2024.State.Legislative.Districts...Upper.NAME
#> 1                        State Senate District 5
#>   X2024.State.Legislative.Districts...Upper.OBJECTID
#> 1                                                796
#>   X2024.State.Legislative.Districts...Upper.CENTLON
#> 1                                      -090.2277047
#>   X2024.State.Legislative.Districts...Upper.LSY
#> 1                                          2024
#>   X2024.State.Legislative.Districts...Upper.AREALAND
#> 1                                          105540123
#>   X2024.State.Legislative.Districts...Upper.INTPTLON
#> 1                                       -090.2284231
#>   X2024.State.Legislative.Districts...Upper.MTFCC
#> 1                                           G5210
#>   X2024.State.Legislative.Districts...Upper.LDTYP
#> 1                                               O
#>   X2024.State.Legislative.Districts...Lower.GEOID
#> 1                                           29079
#>   X2024.State.Legislative.Districts...Lower.CENTLAT
#> 1                                       +38.6369287
#>   X2024.State.Legislative.Districts...Lower.SLDL
#> 1                                            079
#>   X2024.State.Legislative.Districts...Lower.AREAWATER
#> 1                                                   0
#>   X2024.State.Legislative.Districts...Lower.STATE
#> 1                                              29
#>   X2024.State.Legislative.Districts...Lower.BASENAME
#> 1                                                 79
#>   X2024.State.Legislative.Districts...Lower.OID
#> 1                              2139035951822122
#>   X2024.State.Legislative.Districts...Lower.LSADC
#> 1                                              LL
#>   X2024.State.Legislative.Districts...Lower.FUNCSTAT
#> 1                                                  N
#>   X2024.State.Legislative.Districts...Lower.INTPTLAT
#> 1                                        +38.6369287
#>   X2024.State.Legislative.Districts...Lower.NAME
#> 1                        State House District 79
#>   X2024.State.Legislative.Districts...Lower.OBJECTID
#> 1                                                966
#>   X2024.State.Legislative.Districts...Lower.CENTLON
#> 1                                      -090.2290250
#>   X2024.State.Legislative.Districts...Lower.LSY
#> 1                                          2024
#>   X2024.State.Legislative.Districts...Lower.AREALAND
#> 1                                           18628738
#>   X2024.State.Legislative.Districts...Lower.INTPTLON
#> 1                                       -090.2290250
#>   X2024.State.Legislative.Districts...Lower.MTFCC
#> 1                                           G5220
#>   X2024.State.Legislative.Districts...Lower.LDTYP X2020.Census.Blocks.SUFFIX
#> 1                                               O                           
#>   X2020.Census.Blocks.GEOID X2020.Census.Blocks.CENTLAT
#> 1           295101278001042                 +38.6362480
#>   X2020.Census.Blocks.BLOCK X2020.Census.Blocks.AREAWATER
#> 1                      1042                             0
#>   X2020.Census.Blocks.STATE X2020.Census.Blocks.BASENAME
#> 1                        29                         1042
#>   X2020.Census.Blocks.OID X2020.Census.Blocks.LSADC
#> 1         210701005649120                        BK
#>   X2020.Census.Blocks.FUNCSTAT X2020.Census.Blocks.INTPTLAT
#> 1                            S                  +38.6362480
#>   X2020.Census.Blocks.NAME X2020.Census.Blocks.OBJECTID
#> 1               Block 1042                      1068723
#>   X2020.Census.Blocks.TRACT X2020.Census.Blocks.CENTLON
#> 1                    127800                -090.2318925
#>   X2020.Census.Blocks.BLKGRP X2020.Census.Blocks.AREALAND
#> 1                          1                        36036
#>   X2020.Census.Blocks.INTPTLON X2020.Census.Blocks.MTFCC
#> 1                 -090.2318925                     G5040
#>   X2020.Census.Blocks.LWBLKTYP X2020.Census.Blocks.UR
#> 1                            L                      U
#>   X2020.Census.Blocks.COUNTY Census.Tracts.GEOID Census.Tracts.CENTLAT
#> 1                        510         29510127800           +38.6319521
#>   Census.Tracts.AREAWATER Census.Tracts.STATE Census.Tracts.BASENAME
#> 1                       0                  29                   1278
#>   Census.Tracts.OID Census.Tracts.LSADC Census.Tracts.FUNCSTAT
#> 1  2079015485646034                  CT                      S
#>   Census.Tracts.INTPTLAT Census.Tracts.NAME Census.Tracts.OBJECTID
#> 1            +38.6319521  Census Tract 1278                  23319
#>   Census.Tracts.TRACT Census.Tracts.CENTLON Census.Tracts.AREALAND
#> 1              127800          -090.2259200                2546610
#>   Census.Tracts.INTPTLON Census.Tracts.MTFCC Census.Tracts.COUNTY
#> 1           -090.2259200               G5020                  510
#>   X119th.Congressional.Districts.GEOID X119th.Congressional.Districts.CENTLAT
#> 1                                 2901                            +38.7249542
#>   X119th.Congressional.Districts.CDSESSN
#> 1                                    119
#>   X119th.Congressional.Districts.AREAWATER
#> 1                                 27393419
#>   X119th.Congressional.Districts.BASENAME X119th.Congressional.Districts.STATE
#> 1                                       1                                   29
#>   X119th.Congressional.Districts.OID X119th.Congressional.Districts.LSADC
#> 1                   2119035951822595                                   C2
#>   X119th.Congressional.Districts.FUNCSTAT
#> 1                                       N
#>   X119th.Congressional.Districts.INTPTLAT X119th.Congressional.Districts.NAME
#> 1                             +38.7269305            Congressional District 1
#>   X119th.Congressional.Districts.OBJECTID
#> 1                                     139
#>   X119th.Congressional.Districts.CENTLON X119th.Congressional.Districts.CD119
#> 1                           -090.3057680                                   01
#>   X119th.Congressional.Districts.AREALAND
#> 1                               642192503
#>   X119th.Congressional.Districts.INTPTLON X119th.Congressional.Districts.MTFCC
#> 1                            -090.3058980                                G5200
# }