Geocode Single Structured Address

Provides access to the structured single address geocoding API from the US Census Bureau.

cxy_single(
  street,
  city = NULL,
  state = NULL,
  zip = NULL,
  return = "locations",
  benchmark = "Public_AR_Current",
  vintage = NULL
)

Arguments

street
String containing street address

city
Optional String containing city

state
Optional String containing state

zip
Optional String or Integer containing 5-digit Zip Code

return
One of 'locations' or 'geographies' See Details.

benchmark
Optional ID or Name of Census Benchmark. See Details.

vintage
Optional ID or Name of Census Vintage. See Details.

Value

A data.frame containing matched address or NULL if not matches

Details

To obtain current valid benchmarks, use the cxy_benchmarks() function

If you want to append census geographies, you must specify a valid vintage for your benchmark. You may use the cxy_vintages() function to obtain valid Vintages. See vignette('censusxy') for a full walkthrough.