Geocode Single Coordinate Pair

Provides access to the GeoLookup API of the US Census Bureau. Returns census geographies for a single geographic point.

cxy_geography(
  lon,
  lat,
  benchmark = "Public_AR_Current",
  vintage = "Current_Current"
)

Arguments

lon

Numeric or String Containing Longitude (x) of Point

lat

Numeric or String Containing Latitude (y) of Point

benchmark

Optional ID or Name of Census Benchmark. See Details.

vintage

Optional ID or Name of Census Vintage. See Details.

Value

A data.frame containing matched address or NULL if not matches

Details

This function can be used to locate geographic information given a geographic point. It does not provide an address like a reverse-geocoder

To obtain current valid benchmarks, use the cxy_benchmarks() function

To use this function, you must specify a valid vintage for your benchmark. You may use the cxy_vintages() function to obtain valid Vintages for a given benchmark. See vignette('censusxy') for a full walkthrough.

Examples

cxy_geography(lon = -90.23324, lat = 38.63593)
#>   X2018.State.Legislative.Districts...Upper.GEOID
#> 1                                           29005
#>   X2018.State.Legislative.Districts...Upper.CENTLAT
#> 1                                       +38.6292642
#>   X2018.State.Legislative.Districts...Upper.AREAWATER
#> 1                                             7648466
#>   X2018.State.Legislative.Districts...Upper.STATE
#> 1                                              29
#>   X2018.State.Legislative.Districts...Upper.BASENAME
#> 1                                                  5
#>   X2018.State.Legislative.Districts...Upper.OID
#> 1                               212904690745019
#>   X2018.State.Legislative.Districts...Upper.SLDU
#> 1                                            005
#>   X2018.State.Legislative.Districts...Upper.LSADC
#> 1                                              LU
#>   X2018.State.Legislative.Districts...Upper.FUNCSTAT
#> 1                                                  N
#>   X2018.State.Legislative.Districts...Upper.INTPTLAT
#> 1                                        +38.6289053
#>   X2018.State.Legislative.Districts...Upper.NAME
#> 1                        State Senate District 5
#>   X2018.State.Legislative.Districts...Upper.OBJECTID
#> 1                                               1584
#>   X2018.State.Legislative.Districts...Upper.CENTLON
#> 1                                      -090.2261035
#>   X2018.State.Legislative.Districts...Upper.LSY
#> 1                                          2018
#>   X2018.State.Legislative.Districts...Upper.AREALAND
#> 1                                           87740306
#>   X2018.State.Legislative.Districts...Upper.INTPTLON
#> 1                                       -090.2271308
#>   X2018.State.Legislative.Districts...Upper.MTFCC
#> 1                                           G5210
#>   X2018.State.Legislative.Districts...Upper.LDTYP States.STATENS States.GEOID
#> 1                                               O       01779791           29
#>   States.CENTLAT States.AREAWATER States.STATE States.BASENAME States.STUSAB
#> 1    +38.3569829       2487526202           29        Missouri            MO
#>      States.OID States.LSADC States.FUNCSTAT States.INTPTLAT States.DIVISION
#> 1 2749046934061           00               A     +38.3507500               4
#>   States.NAME States.REGION States.OBJECTID States.CENTLON States.AREALAND
#> 1    Missouri             2              50   -092.4580564    178052253239
#>   States.INTPTLON States.MTFCC County.Subdivisions.COUSUB
#> 1    -092.4567826        G4000                      65000
#>   County.Subdivisions.GEOID County.Subdivisions.CENTLAT
#> 1                2951065000                 +38.6357916
#>   County.Subdivisions.AREAWATER County.Subdivisions.STATE
#> 1                      11534970                        29
#>   County.Subdivisions.BASENAME County.Subdivisions.OID
#> 1                    St. Louis          27690325436609
#>   County.Subdivisions.LSADC County.Subdivisions.FUNCSTAT
#> 1                        25                            F
#>   County.Subdivisions.INTPTLAT County.Subdivisions.NAME
#> 1                  +38.6356988           St. Louis city
#>   County.Subdivisions.OBJECTID County.Subdivisions.CENTLON
#> 1                        20365                -090.2451153
#>   County.Subdivisions.COUSUBCC County.Subdivisions.AREALAND
#> 1                           Z7                    159853177
#>   County.Subdivisions.INTPTLON County.Subdivisions.MTFCC
#> 1                 -090.2445816                     G4040
#>   County.Subdivisions.COUSUBNS County.Subdivisions.COUNTY
#> 1                     00767557                        510
#>   Incorporated.Places.NECTAPCI Incorporated.Places.GEOID
#> 1                            N                   2965000
#>   Incorporated.Places.CENTLAT Incorporated.Places.AREAWATER
#> 1                 +38.6357916                      11534970
#>   Incorporated.Places.STATE Incorporated.Places.BASENAME
#> 1                        29                    St. Louis
#>   Incorporated.Places.OID Incorporated.Places.LSADC Incorporated.Places.PLACE
#> 1          27890325442361                        25                     65000
#>   Incorporated.Places.FUNCSTAT Incorporated.Places.INTPTLAT
#> 1                            A                  +38.6356988
#>   Incorporated.Places.NAME Incorporated.Places.OBJECTID
#> 1           St. Louis city                         9057
#>   Incorporated.Places.PLACECC Incorporated.Places.CENTLON
#> 1                          C7                -090.2451153
#>   Incorporated.Places.CBSAPCI Incorporated.Places.AREALAND
#> 1                           N                    159853177
#>   Incorporated.Places.INTPTLON Incorporated.Places.PLACENS
#> 1                 -090.2445816                    00767557
#>   Incorporated.Places.MTFCC Counties.GEOID Counties.CENTLAT Counties.AREAWATER
#> 1                     G4110          29510      +38.6357916           11534970
#>   Counties.STATE Counties.BASENAME   Counties.OID Counties.LSADC
#> 1             29         St. Louis 27590325427938             25
#>   Counties.FUNCSTAT Counties.INTPTLAT  Counties.NAME Counties.OBJECTID
#> 1                 F       +38.6356988 St. Louis city              1817
#>   Counties.CENTLON Counties.COUNTYCC Counties.COUNTYNS Counties.AREALAND
#> 1     -090.2451153                C7          00767557         159853177
#>   Counties.INTPTLON Counties.MTFCC Counties.COUNTY
#> 1      -090.2445816          G4020             510
#>   X2018.State.Legislative.Districts...Lower.GEOID
#> 1                                           29079
#>   X2018.State.Legislative.Districts...Lower.CENTLAT
#> 1                                       +38.6331702
#>   X2018.State.Legislative.Districts...Lower.SLDL
#> 1                                            079
#>   X2018.State.Legislative.Districts...Lower.AREAWATER
#> 1                                                   0
#>   X2018.State.Legislative.Districts...Lower.STATE
#> 1                                              29
#>   X2018.State.Legislative.Districts...Lower.BASENAME
#> 1                                                 79
#>   X2018.State.Legislative.Districts...Lower.OID
#> 1                               213904690745049
#>   X2018.State.Legislative.Districts...Lower.LSADC
#> 1                                              LL
#>   X2018.State.Legislative.Districts...Lower.FUNCSTAT
#> 1                                                  N
#>   X2018.State.Legislative.Districts...Lower.INTPTLAT
#> 1                                        +38.6331702
#>   X2018.State.Legislative.Districts...Lower.NAME
#> 1                        State House District 79
#>   X2018.State.Legislative.Districts...Lower.OBJECTID
#> 1                                               3495
#>   X2018.State.Legislative.Districts...Lower.CENTLON
#> 1                                      -090.2247821
#>   X2018.State.Legislative.Districts...Lower.LSY
#> 1                                          2018
#>   X2018.State.Legislative.Districts...Lower.AREALAND
#> 1                                           21257671
#>   X2018.State.Legislative.Districts...Lower.INTPTLON
#> 1                                       -090.2247821
#>   X2018.State.Legislative.Districts...Lower.MTFCC
#> 1                                           G5220
#>   X2018.State.Legislative.Districts...Lower.LDTYP
#> 1                                               O
#>   X116th.Congressional.Districts.GEOID X116th.Congressional.Districts.CENTLAT
#> 1                                 2901                            +38.7274717
#>   X116th.Congressional.Districts.CDSESSN
#> 1                                    116
#>   X116th.Congressional.Districts.AREAWATER X116th.Congressional.Districts.STATE
#> 1                                 27418730                                   29
#>   X116th.Congressional.Districts.BASENAME X116th.Congressional.Districts.OID
#> 1                                       1                    211904690723734
#>   X116th.Congressional.Districts.LSADC X116th.Congressional.Districts.FUNCSTAT
#> 1                                   C2                                       N
#>   X116th.Congressional.Districts.INTPTLAT X116th.Congressional.Districts.NAME
#> 1                             +38.7285909            Congressional District 1
#>   X116th.Congressional.Districts.OBJECTID
#> 1                                     236
#>   X116th.Congressional.Districts.CENTLON
#> 1                           -090.2944715
#>   X116th.Congressional.Districts.AREALAND
#> 1                               583673550
#>   X116th.Congressional.Districts.INTPTLON X116th.Congressional.Districts.CD116
#> 1                            -090.2965080                                   01
#>   X116th.Congressional.Districts.MTFCC X2020.Census.Blocks.SUFFIX
#> 1                                G5200                           
#>   X2020.Census.Blocks.GEOID X2020.Census.Blocks.CENTLAT
#> 1           295101278001042                 +38.6362480
#>   X2020.Census.Blocks.BLOCK X2020.Census.Blocks.AREAWATER
#> 1                      1042                             0
#>   X2020.Census.Blocks.STATE X2020.Census.Blocks.BASENAME
#> 1                        29                         1042
#>   X2020.Census.Blocks.OID X2020.Census.Blocks.LSADC
#> 1         210701005649120                        BK
#>   X2020.Census.Blocks.FUNCSTAT X2020.Census.Blocks.INTPTLAT
#> 1                            S                  +38.6362480
#>   X2020.Census.Blocks.NAME X2020.Census.Blocks.OBJECTID
#> 1               Block 1042                      2255916
#>   X2020.Census.Blocks.TRACT X2020.Census.Blocks.CENTLON
#> 1                    127800                -090.2318925
#>   X2020.Census.Blocks.BLKGRP X2020.Census.Blocks.AREALAND
#> 1                          1                        36036
#>   X2020.Census.Blocks.INTPTLON X2020.Census.Blocks.MTFCC
#> 1                 -090.2318925                     G5040
#>   X2020.Census.Blocks.LWBLKTYP X2020.Census.Blocks.COUNTY Census.Tracts.GEOID
#> 1                            L                        510         29510127800
#>   Census.Tracts.CENTLAT Census.Tracts.AREAWATER Census.Tracts.STATE
#> 1           +38.6319521                       0                  29
#>   Census.Tracts.BASENAME Census.Tracts.OID Census.Tracts.LSADC
#> 1                   1278  2079015485646034                  CT
#>   Census.Tracts.FUNCSTAT Census.Tracts.INTPTLAT Census.Tracts.NAME
#> 1                      S            +38.6319521  Census Tract 1278
#>   Census.Tracts.OBJECTID Census.Tracts.TRACT Census.Tracts.CENTLON
#> 1                  23228              127800          -090.2259200
#>   Census.Tracts.AREALAND Census.Tracts.INTPTLON Census.Tracts.MTFCC
#> 1                2546610           -090.2259200               G5020
#>   Census.Tracts.COUNTY
#> 1                  510