Geocode Single Coordinate Pair

Provides access to the GeoLookup API of the US Census Bureau. Returns census geographies for a single geographic point.

cxy_geography(
  lon,
  lat,
  benchmark = "Public_AR_Current",
  vintage = "Current_Current"
)

Arguments

lon
Numeric or String Containing Longitude (x) of Point

lat
Numeric or String Containing Latitude (y) of Point

benchmark
Optional ID or Name of Census Benchmark. See Details.

vintage
Optional ID or Name of Census Vintage. See Details.

Value

A data.frame containing matched address or NULL if not matches

Details

This function can be used to locate geographic information given a geographic point. It does not provide an address like a reverse-geocoder

To obtain current valid benchmarks, use the cxy_benchmarks() function

To use this function, you must specify a valid vintage for your benchmark. You may use the cxy_vintages() function to obtain valid Vintages for a given benchmark. See vignette('censusxy') for a full walkthrough.

Examples

cxy_geography(lon = -90.23324, lat = 38.63593)

#>   States.STATENS States.GEOID States.CENTLAT States.AREAWATER States.STATE
#> 1       01779791           29    +38.3569829       2487215790           29
#>   States.BASENAME States.STUSAB   States.OID States.LSADC States.FUNCSTAT
#> 1        Missouri            MO 2.749047e+12           00               A
#>   States.INTPTLAT States.DIVISION States.NAME States.REGION States.OBJECTID
#> 1     +38.3507500               4    Missouri             2              46
#>   States.CENTLON States.AREALAND States.INTPTLON States.MTFCC Counties.GEOID
#> 1   -092.4580564    178052563675    -092.4567826        G4000          29510
#>   Counties.CENTLAT Counties.AREAWATER Counties.STATE Counties.BASENAME
#> 1      +38.6358185           11563208             29         St. Louis
#>   Counties.OID Counties.LSADC Counties.FUNCSTAT Counties.INTPTLAT
#> 1 2.759033e+13             25                 F       +38.6356988
#>    Counties.NAME Counties.OBJECTID Counties.CENTLON Counties.COUNTYCC
#> 1 St. Louis city              3142     -090.2451165                C7
#>   Counties.COUNTYNS Counties.AREALAND Counties.INTPTLON Counties.MTFCC
#> 1          00767557         159895650      -090.2445816          G4020
#>   Counties.COUNTY X2020.Census.Blocks.SUFFIX X2020.Census.Blocks.GEOID
#> 1             510                                      295101278001042
#>   X2020.Census.Blocks.CENTLAT X2020.Census.Blocks.BLOCK
#> 1                 +38.6362480                      1042
#>   X2020.Census.Blocks.AREAWATER X2020.Census.Blocks.STATE
#> 1                             0                        29
#>   X2020.Census.Blocks.BASENAME X2020.Census.Blocks.OID
#> 1                         1042             2.10701e+14
#>   X2020.Census.Blocks.LSADC X2020.Census.Blocks.FUNCSTAT
#> 1                        BK                            S
#>   X2020.Census.Blocks.INTPTLAT X2020.Census.Blocks.NAME
#> 1                  +38.6362480               Block 1042
#>   X2020.Census.Blocks.OBJECTID X2020.Census.Blocks.TRACT
#> 1                      4862904                    127800
#>   X2020.Census.Blocks.CENTLON X2020.Census.Blocks.BLKGRP
#> 1                -090.2318925                          1
#>   X2020.Census.Blocks.AREALAND X2020.Census.Blocks.INTPTLON
#> 1                        36036                 -090.2318925
#>   X2020.Census.Blocks.MTFCC X2020.Census.Blocks.LWBLKTYP
#> 1                     G5040                            L
#>   X2020.Census.Blocks.COUNTY Census.Tracts.GEOID Census.Tracts.CENTLAT
#> 1                        510         29510127800           +38.6319521
#>   Census.Tracts.AREAWATER Census.Tracts.STATE Census.Tracts.BASENAME
#> 1                       0                  29                   1278
#>   Census.Tracts.OID Census.Tracts.LSADC Census.Tracts.FUNCSTAT
#> 1      2.079015e+15                  CT                      S
#>   Census.Tracts.INTPTLAT Census.Tracts.NAME Census.Tracts.OBJECTID
#> 1            +38.6319521  Census Tract 1278                  23417
#>   Census.Tracts.TRACT Census.Tracts.CENTLON Census.Tracts.AREALAND
#> 1              127800          -090.2259200                2546610
#>   Census.Tracts.INTPTLON Census.Tracts.MTFCC Census.Tracts.COUNTY
#> 1           -090.2259200               G5020                  510