Geocode Single Coordinate Pair

Provides access to the GeoLookup API of the US Census Bureau. Returns census geographies for a single geographic point.

cxy_geography(
  lon,
  lat,
  benchmark = "Public_AR_Current",
  vintage = "Current_Current"
)

Arguments

lon

Numeric or String Containing Longitude (x) of Point

lat

Numeric or String Containing Latitude (y) of Point

benchmark

Optional ID or Name of Census Benchmark. See Details.

vintage

Optional ID or Name of Census Vintage. See Details.

Value

A data.frame containing matched address or NULL if not matches

Details

This function can be used to locate geographic information given a geographic point. It does not provide an address like a reverse-geocoder

To obtain current valid benchmarks, use the cxy_benchmarks() function

To use this function, you must specify a valid vintage for your benchmark. You may use the cxy_vintages() function to obtain valid Vintages for a given benchmark. See vignette('censusxy') for a full walkthrough.

Examples

# \donttest{
cxy_geography(lon = -90.23324, lat = 38.63593)
#> Error in cxy_geography(lon = -90.23324, lat = 38.63593): The server, while working as a gateway to get a response needed to handle the request, got an invalid response.
# }