Homicides in the City of St. Louis, 2008 - 2018

An example data set containing the addresses for Homicides reported by the Saint Louis Metropolitan Police Department

data(stl_homicides)

Format

A tibble with 1822 rows and 6 variables:

street_address

number, street and street suffix where homicide occured

year

year homicide occured

date

data homicide occured

state

state abbreviation of location, in these data, all "MO"

postal_code

zipcode/postal code of location, in these data all NA

city

city of location, in these data all "St. Louis"

Source

St. Louis Metropolitan Police Department

Examples

str(stl_homicides)

#> Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	1822 obs. of  6 variables:
#>  $ street_address: chr  "5738 Terry Ave" "5356 Page Blvd" "5826 Roosevelt Pl" "3859 Ohio Ave" ...
#>  $ year          : num  2008 2008 2008 2008 2008 ...
#>  $ date          : chr  "01/12/2008 12:37" "01/17/2008 04:00" "01/20/2008 21:19" "01/21/2008 17:38" ...
#>  $ state         : chr  "MO" "MO" "MO" "MO" ...
#>  $ postal_code   : logi  NA NA NA NA NA NA ...
#>  $ city          : chr  "St. Louis" "St. Louis" "St. Louis" "St. Louis" ...

head(stl_homicides)

#> # A tibble: 6 x 6
#>   street_address        year date             state postal_code city     
#>   <chr>                <dbl> <chr>            <chr> <lgl>       <chr>    
#> 1 5738 Terry Ave        2008 01/12/2008 12:37 MO    NA          St. Louis
#> 2 5356 Page Blvd        2008 01/17/2008 04:00 MO    NA          St. Louis
#> 3 5826 Roosevelt Pl     2008 01/20/2008 21:19 MO    NA          St. Louis
#> 4 3859 Ohio Ave         2008 01/21/2008 17:38 MO    NA          St. Louis
#> 5 4100 Saint Louis Ave  2008 01/30/2008 15:34 MO    NA          St. Louis
#> 6 2418 N Euclid Ave     2008 01/30/2008 19:19 MO    NA          St. Louis